+++ TaskManagement.Application/DTOs/RequestDTOs/Organization/LeaveUserOrgAppDto.cs
@@ -1,5 +1,5 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Organization;
 public record LeaveUserOrgAppDto(
-    int UserId,
-    int OrgId
+    long UserId,
+    long OrgId
 );
