+++ TaskManagement.Application/DTOs/RequestDTOs/Project/AddRemoveUserProjectAppDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Project;
 public record AddRemoveUserProjectAppDto(
-    int UserId,
-    int ProjId,
-    int OwnerId
+    long UserId,
+    long ProjId,
+    long OwnerId
 );
