+++ TaskManagement.Application/DTOs/RequestDTOs/Project/ChangeProjectProgressAppDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Project;
 public record ChangeProjectProgressAppDto(
-    int OwnerId,
-    int ProjId,
+    long OwnerId,
+    long ProjId,
     byte ProjectProgress
 );
