+++ TaskManagement.Application/DTOs/RequestDTOs/Task/AddRemoveUserTaskAppDto.cs
@@ -1,7 +1,7 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Task;
 public record AddRemoveUserTaskAppDto(
-    int OwnerId,
-    int UserId,
-    int TaskId,
-    int ProjId
+    long OwnerId,
+    long UserId,
+    long TaskId,
+    long ProjId
 );
