+++ TaskManagement.Application/DTOs/RequestDTOs/Task/ChangeTaskActivityAppDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Task;
 public record ChangeTaskActivityAppDto(
-    int UserId,
-    int TaskId,
+    long UserId,
+    long TaskId,
     bool Activity
 );
