+++ TaskManagement.Application/Features/Invitation/Command/AcceptInvitation/AcceptInvitationCommand.cs
@@ -3,6 +3,6 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Invitation.Command.AcceptInvitation;
 public record AcceptInvitationCommand(
-    int UserId,
+    long UserId,
     string Token
 ) : IRequest<GeneralResult>;
