+++ TaskManagement.Application/Features/Invitation/Command/GenerateInviteLinkByUserId/GenerateInviteLinkByUserIdCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Invitation.Command.GenerateInviteLinkByUserId;
 public record GenerateInviteLinkByUserIdCommand(
-    int OrgId,
-    int OrgOwnerId,
+    long OrgId,
+    long OrgOwnerId,
     string UserMobileNumber
 ) : IRequest<GeneralResult<string>>;
