+++ TaskManagement.Application/Features/Organization/Command/ChangeOrgActivity/ChangeOrgActivityCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Organization.Command.ChangeOrgActivity;
 public record ChangeOrgActivityCommand(
-    int OrgId,
-    int OwnerId,
+    long OrgId,
+    long OwnerId,
     string OwnerPassword
 ) : IRequest<GeneralResult>;
