+++ TaskManagement.Application/Features/Organization/Command/ChangeUserRoleToAdmin/ChangeUserRoleToAdminCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Organization.Command.ChangeUserRoleToAdmin;
 public record ChangeUserRoleToAdminCommand(
-    int OrgOwnerId,
-    int OrgId,
-    int UserId
+    long OrgOwnerId,
+    long OrgId,
+    long UserId
 ) : IRequest<GeneralResult>;
