+++ TaskManagement.Application/Features/Organization/Command/DeleteOrg/DeleteOrgCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Organization.Command.DeleteOrg;
 public record DeleteOrgCommand(
-    int OrgId,
-    int OwnerId,
+    long OrgId,
+    long OwnerId,
     string OwnerPassword
 ) : IRequest<GeneralResult>;
