+++ TaskManagement.Application/Features/Organization/Query/GetOrgById/GetOrgByIdQuery.cs
@@ -3,5 +3,5 @@ using TaskManagement.Application.DTOs.ResponseDTOs.Organization;
 using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Organization.Query.GetOrgById;
-public record GetOrgByIdQuery(int OrgId)
+public record GetOrgByIdQuery(long OrgId)
     : IRequest<GeneralResult<OrgDetailsDto>>;
