+++ TaskManagement.Application/Features/Project/Command/AddUserToProject/AddUserToProjectCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Project.Command.AddUserToProject;
 public record AddUserToProjectCommand(
-    int UserId,
-    int ProjId,
-    int OwnerId
+    long UserId,
+    long ProjId,
+    long OwnerId
 ) : IRequest<GeneralResult>;
