+++ TaskManagement.Application/Features/Project/Command/ChangeProjectActivity/ChangeProjectActivityCommand.cs
@@ -3,8 +3,8 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Project.Command.ChangeProjectActivity;
 public record ChangeProjectActivityCommand(
-    int OwnerId,
-    int ProjId,
+    long OwnerId,
+    long ProjId,
     string UserPassword,
     bool Activity
 ) : IRequest<GeneralResult>;
