+++ TaskManagement.Application/Features/Project/Command/ChangeProjectProgress/ChangeProjectProgressCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Project.Command.ChangeProjectProgress;
 public record ChangeProjectProgressCommand(
-    int OwnerId,
-    int ProjId,
+    long OwnerId,
+    long ProjId,
     byte ProjectProgress
 ) : IRequest<GeneralResult>;
