+++ TaskManagement.Application/Features/Project/Command/ChangeUserRoleToAdmin/ChangeUserRoleToAdminCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Project.Command.ChangeUserRoleToAdmin;
 public record ChangeUserRoleToAdminCommand(
-    int OwnerId,
-    int ProjId,
-    int UserId
+    long OwnerId,
+    long ProjId,
+    long UserId
 ) : IRequest<GeneralResult>;
