+++ TaskManagement.Application/Features/Project/Command/ChangeUserRoleToMember/ChangeUserRoleToMemberCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Project.Command.ChangeUserRoleToMember;
 public record ChangeUserRoleToMemberCommand(
-    int OwnerId,
-    int ProjId,
-    int UserId
+    long OwnerId,
+    long ProjId,
+    long UserId
 ) : IRequest<GeneralResult>;
