+++ TaskManagement.Application/Features/Project/Query/GetProjectById/GetProjectByIdQuery.cs
@@ -3,5 +3,5 @@ using TaskManagement.Application.DTOs.ResponseDTOs.Project;
 using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Project.Query.GetProjectById;
-public record GetProjectByIdQuery(int ProjectId)
+public record GetProjectByIdQuery(long ProjectId)
     : IRequest<GeneralResult<ProjectDetailsDto>>;
