+++ TaskManagement.Application/Features/Task/Command/StartTask/StartTaskCommand.cs
@@ -3,6 +3,6 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Task.Command.StartTask;
 public record StartTaskCommand(
-    int UserId,
-    int TaskId
+    long UserId,
+    long TaskId
 ) : IRequest<GeneralResult>;
