+++ TaskManagement.Application/Features/Task/Query/GetTaskByIdQuery.cs
@@ -3,5 +3,5 @@ using TaskManagement.Application.DTOs.ResponseDTOs.Task;
 using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Task.Query;
-public record GetTaskByIdQuery(int TaskId)
+public record GetTaskByIdQuery(long TaskId)
     : IRequest<GeneralResult<TaskDetailsDto>>;
